vbe: Add a device for VBE simple
Add the 'simple' VBE driver which supports updating a single firmware
image in MMC, such as is used by ROCKPro64 and many other boards.

The implementation reads the FIT out of the .cab file, selects the
best configuration to use based optionally on the model name, then writes
all the images in that configuration.

Note that this is mostly designed for platforms which use device tree,
such as ARM. For those that use ACPI, a system device tree is supported
so that things work. Future work may introduce another way of providing
the details needed by VBE.

Signed-off-by: Simon Glass <sjg@chromium.org>