vbe: Add an example showing how to build an update
A VBE update consists of a FIT embedded in a .cab file, with compression
provided by the latter, at least for now.

Add an example script to build an update which works with VBE simple.
Also add sample device trees for testing on the host and with a ROCKPro64
board.

Signed-off-by: Simon Glass <sjg@chromium.org>